+++ continual/__init__.py
@@ -1,4 +1,5 @@
-from .container import Parallel, Residual, Sequential  # noqa: F401
+from .closure import Add, Lambda, Multiply, Unity  # noqa: F401
+from .container import Conditional, Parallel, Residual, Sequential  # noqa: F401
 from .conv import Conv1d, Conv2d, Conv3d  # noqa: F401
 from .convert import continual, forward_stepping  # noqa: F401
 from .delay import Delay  # noqa: F401