+++ CHANGELOG.md
@@ -10,6 +10,15 @@ From v1.0.0 and on, the project will adherence strictly to Semantic Versioning.
 ## [Unreleased]
 
 
+## [0.9.0]
+### Added
+- `co.Lambda` module.
+- `co.Add` module.
+- `co.Multiply` module.
+- `co.Unity` module.
+- `co.Conditional` module.
+
+
 ## [0.8.1]
 ### Fixed
 - Bug in `forward_stepping`.